djangoproject/requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cfgv==3.5.0
distlib==0.4.0
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
filelock==3.25.2
gunicorn==25.1.0
identify==2.6.18
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
nodeenv==1.10.0
packaging==26.0
platformdirs==4.9.4
pluggy==1.6.0
pre_commit==4.5.1
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.12.0
python-discovery==1.1.3
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
ruff==0.15.6
sentry-sdk==2.54.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==21.2.0
whitenoise==6.12.0
